% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-server.R
\name{app_server}
\alias{app_server}
\title{App server}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}
}
\value{
none
}
\description{
Create the server-side component of the dccvalidator Shiny app.
}
\examples{
\dontrun{
shinyApp(ui = app_ui, server = app_server)
}
}
