% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-ui.R
\name{app_ui}
\alias{app_ui}
\title{App UI}
\usage{
app_ui(request)
}
\arguments{
\item{request}{Shiny request object}
}
\value{
A shinydashboard page
}
\description{
If not running interactively (i.e. locally), launches the
\code{\link{mod_synapse_oauth_ui}} to start dccvalidator using OAuth for
login. Otherwise, launches the \code{\link{mod_main_ui}} to start
dccvalidator using login credentials stored in a .synapseConfig.
}
\examples{
\dontrun{
shinyApp(ui = app_ui, server = app_server)
}
}
