% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse-helpers.R
\name{attempt_login}
\alias{attempt_login}
\title{Attempt to log into Synapse}
\usage{
attempt_login(syn, ...)
}
\arguments{
\item{syn}{Synapse client object}

\item{...}{Synapse credentials, such as \code{authToken} or \code{email} with a
\code{password} or \code{apiKey}.}
}
\value{
TRUE if logged in, else \code{NULL}
}
\description{
Attempt to log into Synapse. Will first try using authentication
credentials written to a .synapseConfig file. If that fails, will try using
any credentials passed to the function. Will return \code{NULL} if not all
attempts failed.
}
