% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-annotation-values.R
\name{can_coerce}
\alias{can_coerce}
\title{Check coercibility}
\usage{
can_coerce(values, class)
}
\arguments{
\item{values}{Vector of values to check}

\item{class}{Class of interest}
}
\value{
Boolean value; TRUE if \code{values} are coercible to \code{class}, \code{FALSE}
otherwise.
}
\description{
Checks if values are coercible to a given class. Because of inconsistencies
in R's built-in coercion functions (e.g. \code{as.numeric()} warns when it
introduces NAs but \code{as.logical()} doesn't; \code{as.integer()} will silently
remove decimal places from numeric inputs) we check only for the specific
coercions we want to allow, currently: numeric, integer, or logical to string
}
\details{
This function is mainly in place so that we can automatically allow numeric
read lengths, pH values, etc., which are defined as strings in our annotation
vocabulary but can reasonably be numbers.

This function will \emph{not} affect validation of enumerated values, regardless
of their class. It is only used by the \code{\link[=check_type]{check_type()}} function, which
validates annotations that have a required type but no enumerated values.
}
