% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-annotation-keys.R
\name{check_annotation_keys}
\alias{check_annotation_keys}
\title{Check annotation keys}
\usage{
check_annotation_keys(x, annotations, whitelist_keys = NULL)
}
\arguments{
\item{x}{An object to check.}

\item{annotations}{A data frame of annotation definitions. Must contain at
least three columns: \code{key}, \code{value}, and \code{columnType}.}

\item{whitelist_keys}{A character vector keys to whitelist. If these keys are
present in \code{x} but absent from \code{annotations}, they will still be treated as
valid.}
}
\value{
A condition object indicating whether keys match the given annotation
dictionary. Erroneous keys are included as data within the object.
}
\description{
Checks that all annotation keys on a file, in a file view, or in a data frame
are valid annotations. \code{check_annotation_keys()} returns any invalid
annotation keys; \code{valid_annotation_keys()} returns \emph{valid} annotation keys.
}
\examples{
\dontrun{
library("synapser")
synLogin()
annots <- get_synapse_annotations()
my_file <- synGet("syn11931757", downloadFile = FALSE)
check_annotation_keys(my_file, annots)

dat <- data.frame(non_annotation = 5, assay = "rnaSeq")
check_annotation_keys(dat, annots)

fv <- synTableQuery("SELECT * FROM syn17020234")
check_annotation_keys(fv, annots)

# If you don't specify an annotations data frame, these functions will
# download annotations automatically using `get_synapse_annotations()` (must
# be logged in to Synapse)
my_file <- synGet("syn11931757", downloadFile = FALSE)
check_annotation_keys(my_file)
}
}
\seealso{
\code{\link[=valid_annotation_keys]{valid_annotation_keys()}}
}
