% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-annotation-values.R
\name{check_annotation_values}
\alias{check_annotation_values}
\alias{valid_annotation_values}
\title{Check annotation values}
\usage{
check_annotation_values(x, annotations)

valid_annotation_values(x, annotations)
}
\arguments{
\item{x}{An object to check.}

\item{annotations}{A data frame of annotation definitions. Must contain at
least two columns: \code{key} and \code{value}.}
}
\value{
A named list of invalid annotation values.
}
\description{
Checks that all annotation values are valid. Currently only checks enumerated
values for standard annotation keys; it does not check types for annotations
such as \code{compoundDoseUnit} which are required to be strings but have no
enumerated list of values. It also does not report on values for invalid
\emph{keys}; see \code{\link[=check_annotation_keys]{check_annotation_keys()}}.
}
\examples{
\dontrun{
library("synapser")
synLogin()
annots <- get_synapse_annotations()
my_file <- synGet("syn11931757", downloadFile = FALSE)
check_annotation_values(my_file, annots)

dat <- data.frame(non_annotation = 5, assay = "rnaSeq")
check_annotation_values(dat, annots)

fv <- synTableQuery("SELECT * FROM syn17020234")
check_annotation_values(fv, annots)

# If you don't specify an annotations data frame, these functions will
# download annotations automatically using `get_synapse_annotations()` (must
# be logged in to Synapse)
my_file <- synGet("syn11931757", downloadFile = FALSE)
check_annotation_values(my_file)
}
}
