% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-certified-user.R
\name{check_certified_user}
\alias{check_certified_user}
\title{Check if user is certified}
\usage{
check_certified_user(id, syn)
}
\arguments{
\item{id}{User ID}

\item{syn}{Synapse client object}
}
\description{
Check if user has completed and passed the Certified User Quiz.
}
\examples{
\dontrun{
library("synapser")
synLogin()
check_certified_user("3384770")
}
}
