% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-ids.R
\name{check_ids}
\alias{check_ids}
\alias{check_indiv_ids}
\alias{check_specimen_ids}
\title{Check ids}
\usage{
check_ids(x, y, idcol = c("individualID", "specimenID"), xname = NULL,
  yname = NULL)

check_indiv_ids(x, y, xname = NULL, yname = NULL)

check_specimen_ids(x, y, xname = NULL, yname = NULL)
}
\arguments{
\item{x, y}{Data frames to compare}

\item{idcol}{Name of column containing ids to compare}

\item{xname, yname}{Names of x and y (to be used in resulting messages)}
}
\value{
A condition object indicating whether IDs match (\code{"check_pass"}) or
not (\code{"check_fail"}). Mismatched IDs are included as data within the
object.
}
\description{
Compare IDs (such as individual IDs or specimen IDs) between two data frames.

Ensure that all individual IDs in two data frames match.

Ensure that all specimen IDS in two data frames match
}
\examples{
a <- data.frame(individualID = LETTERS[1:3])
b <- data.frame(individualID = LETTERS[1:4])
check_specimen_ids(a, b, "individual", "biospecimen")
a <- data.frame(specimenID = LETTERS[1:3])
b <- data.frame(specimenID = LETTERS[1:4])
check_specimen_ids(a, b, "biospecimen", "assay")
}
