% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-invalid-characters.R
\name{check_invalid_characters}
\alias{check_invalid_characters}
\title{Check for non-ascii characters}
\usage{
check_invalid_characters(
  data,
  success_msg = "There are no invalid characters",
  fail_msg = "There is an invalid character in a column"
)
}
\arguments{
\item{data}{Data to check}

\item{success_msg}{Message indicating the check succeeded.}

\item{fail_msg}{Message indicating the check failed.}
}
\value{
A condition object indicating whether the data contains columns with
a non-ascii character.
}
\description{
Check for non-ascii characters in columns.
}
\examples{
dat <- tibble::tibble(
  fails1 = c("study 1", "study&amp;2"),
  succeeds = c("file1.ext", "file2.ext"),
  fails2 = c("foo<0xa0>", "bar")
)
check_invalid_characters(dat)
}
