% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-team-membership.R
\name{check_team_membership}
\alias{check_team_membership}
\title{Check team membership}
\usage{
check_team_membership(team, user)
}
\arguments{
\item{team}{Team ID to check membership in}

\item{user}{Synapser user object (e.g. output from
\code{\link[synapser:synGetUserProfile]{synapser::synGetUserProfile()}})}
}
\description{
Show a modal dialog if a user isn't a member of a given team.
}
