% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-annotation-values.R
\name{check_type}
\alias{check_type}
\alias{check_value}
\alias{check_values}
\title{Check that class of value matches annotation columnType}
\usage{
check_type(value, key, annotations, return_valid = FALSE)

check_value(values, key, annotations, whitelist_keys = NULL,
  return_valid = FALSE)

check_values(x, annotations, whitelist_keys = NULL,
  return_valid = FALSE)
}
\arguments{
\item{value}{The values of an annotation}

\item{key}{An annotation key}

\item{annotations}{A data frame of annotations to check against}

\item{return_valid}{Should the function return valid values? Defaults to
\code{FALSE} (i.e. the function will return invalid values).}

\item{whitelist_keys}{A chracter vector of annotation keys to whitelist. If
provided, all values for the given keys will be treated as valid.}

\item{x}{A data frame of annotation data}
}
\value{
A vector of invalid values (if \code{return_valid = FALSE}; otherwise a
vector of valid values).

A character vector of valid or invalid values

A named list where each element corresponds to a key that contains
invalid values (if \code{return_valid = FLASE}), and the contents of each
element is a vector of invalid values.
}
\description{
Check that class of value matches annotation columnType

Check values for one key

Check a set of keys and their values
}
