% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-footer.R
\name{create_footer}
\alias{create_footer}
\title{Create app footer}
\usage{
create_footer(email)
}
\arguments{
\item{email}{Email address to contact}
}
\value{
A footer tag
}
\description{
The app footer shows a link to the issue tracker and a contact link.
}
\keyword{internal}
