% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dccvalidator-package.R
\docType{package}
\name{dccvalidator-package}
\alias{dccvalidator}
\alias{dccvalidator-package}
\title{dccvalidator: Validate Data and Metadata}
\description{
Performs automatic checks for common data/metadata quality
    issues.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sage-bionetworks.github.io/dccvalidator/}
  \item \url{https://github.com/Sage-Bionetworks/dccvalidator}
}

}
\author{
\strong{Maintainer}: Kara Woo \email{kara.woo@sagebase.org}

Authors:
\itemize{
  \item Nicole Kauer \email{nicole.kauer@sagebase.org}
  \item Kelsey Montgomery \email{kelsey.montgomery@sagebase.org}
}

Other contributors:
\itemize{
  \item Sage Bionetworks [copyright holder]
}

}
\keyword{internal}
