% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df-to-json.R
\name{df_to_json_list}
\alias{df_to_json_list}
\title{Convert data frame to JSON}
\usage{
df_to_json_list(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A list of JSON blobs
}
\description{
Given a data frame, converts each row to a JSON blob and returns the results
in a list, to make it easier to iteratively validate data with
\code{\link[=check_schema_df]{check_schema_df()}}.
}
\examples{
if (requireNamespace("jsonlite", quietly = TRUE)) {
dat <- data.frame(
  x = c(NA, 1L)
)
df_to_json_list(dat)
}
}
\seealso{
check_schema
}
