% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-synapse-oauth.R
\name{.onLoad}
\alias{.onLoad}
\title{}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{default R .onLoad() parameter}

\item{pkgname}{default R .onLoad() parameter}
}
\description{
Bring in the Synapse Python Client and set up global variables
for OAuth at startup. Requires global variables initialized to NULL: synapse,
app, api, scope, app_url, claims_param, authorization_url.
}
