% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-button-indicator.R
\name{error_func}
\alias{error_func}
\title{Error message helper.}
\usage{
error_func(err, button_id)
}
\arguments{
\item{err}{the error}

\item{button_id}{id for the actionButton}
}
\description{
Error message helper.
}
