% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-specimen-ids.R
\name{find_specimen_mismatches}
\alias{find_specimen_mismatches}
\title{Compare specimenIDs between data file annotations and metadata file contents}
\usage{
find_specimen_mismatches(synID, fileID)
}
\arguments{
\item{synID}{One or more Synapse IDs whose annotations contain specimenIDs to
check.}

\item{fileID}{Synapse ID of the metadata file to check.}
}
\value{
A list of specimen IDs missing from the metadata and from the data.
}
\description{
Compare specimenIDs between data file annotations and metadata file contents
}
\examples{
\dontrun{
synIDs <- c("syn18083968", "syn18083969", "syn18083972", "syn18084124")
meta_file <- "syn18084016"
find_specimen_mismatches(synIDs, meta_file)
}
}
