% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-template.R
\name{get_file_schema}
\alias{get_file_schema}
\title{Get JSON schema attributes from a file}
\usage{
get_file_schema(file)
}
\arguments{
\item{file}{Filepath of json schema.}
}
\description{
Read a JSON schema from a file. Similar
to \link{get_synapse_schema}.
}
