% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df-helpers.R
\name{get_metadataType_indices}
\alias{get_metadataType_indices}
\title{Get metadataType indices}
\usage{
get_metadataType_indices(data, meta_types)
}
\arguments{
\item{data}{A tibble or dataframe with the columns:
name, metadataType, species, assay, file_data, template (optional).
The file_data column should be a list column containing
a dataframe with the file data or \code{NULL} if the data
does not exist. \code{data} is expected to have four rows,
one for each metadataType: individual, biospecimen,
assay, manifest. If file_data is \code{NULL} for a given
metadataType, the metadataType should still be
present. The template column should have the synID or Synapse JSON schema
id for the metadata temaplate. If this is \code{NULL} or the column is missing,
then the check will not verify that all expected columns are present in
the metadata (i.e. \code{check_cols} is skipped).}

\item{meta_types}{List of metadataTypes to find.}
}
\value{
List of indices for \code{data} where the metadataType exists;
\code{NULL} if none exist.
}
\description{
Get the metadataType indices for \code{data}.
}
