% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-specimen-ids.R
\name{get_specimenID_data}
\alias{get_specimenID_data}
\title{Get specimenID annotations from one or more Synapse objects}
\usage{
get_specimenID_data(synID)
}
\arguments{
\item{synID}{One or more Synapse IDs whose annotations contain specimenIDs to
check.}
}
\value{
A named character vector of specimenIDs corresponding to the given
synIDs.
}
\description{
Get specimenID annotations from one or more Synapse objects
}
