% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-specimen-ids.R
\name{get_specimenID_metadata}
\alias{get_specimenID_metadata}
\title{Extract specimen IDs from a metadata file}
\usage{
get_specimenID_metadata(fileID)
}
\arguments{
\item{fileID}{Synapse ID of the metadata file to check.}
}
\value{
A character vector of specimen IDs present in the specimenID column
of the file.
}
\description{
Given the Synapse ID for a metadata file, this function downloads the file,
reads in the first sheet, and looks for the specimenID column. It expects the
file to be an .xlsx file (consistent with AMP-AD templates); this may be
generalized to accommodate CSV files and other column names in the future.
}
