% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-get-study-names.R
\name{get_study}
\alias{get_study}
\title{Server function for the get_study module
Get study name based on either:
\itemize{
\item existing study selection box
\item new name for currently non-existing study
}}
\usage{
get_study(input, output, session)
}
\arguments{
\item{input}{the input variables from shiny::callModule}

\item{output}{the output variables from shiny::callModule}

\item{session}{the session from shiny::callModule}
}
\value{
name of the study
}
\description{
Server function for the get_study module
Get study name based on either:
\itemize{
\item existing study selection box
\item new name for currently non-existing study
}
}
