% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-get-study-names.R
\name{get_study_names}
\alias{get_study_names}
\title{Gather studies within consortium}
\usage{
get_study_names(study_table_id)
}
\arguments{
\item{study_table_id}{synapse Id for the consortium study table}
}
\value{
list of studies
}
\description{
Gather studies within consortium
}
