% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-get-study-names.R
\name{get_study_ui}
\alias{get_study_ui}
\title{UI for the get_study module}
\usage{
get_study_ui(id, study_table_id)
}
\arguments{
\item{id}{the id}

\item{study_table_id}{synapse Id for the consortium study table}
}
\value{
html UI for get_study panel
}
\description{
UI for the get_study module
}
