% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-synapse-annotations.R
\name{get_synapse_annotations}
\alias{get_synapse_annotations}
\title{Get Synapse annotations}
\usage{
get_synapse_annotations(synID = "syn10242922", syn)
}
\arguments{
\item{synID}{The Synapse ID of a table to query from. Defaults to
"syn10242922"}

\item{syn}{Synapse client object}
}
\value{
A data frame containing all annotation keys, descriptions, column
types, maximum sizes, values, value descriptions, sources, and the name of
the annotation's parent module.
}
\description{
Download current annotation values from Synapse and provide them as a data
frame.
}
\examples{
\dontrun{
syn <- synapse$Synapse()
syn$login()
get_synapse_annotations(synID = "syn10242922", syn = syn)
}
}
