% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-template.R
\name{get_synapse_schema}
\alias{get_synapse_schema}
\title{Get Synapse validation schema}
\usage{
get_synapse_schema(syn, id)
}
\arguments{
\item{syn}{Synapse client object}

\item{id}{The id of JSON metadata schema registered in Synapse or Synapse ID
of an excel or csv file containing a metadata template}
}
\description{
Get Synapse validation schema.
}
