% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-synapse-annotations.R
\name{get_synapse_table}
\alias{get_synapse_table}
\title{Get Synapse table}
\usage{
get_synapse_table(synID, syn)
}
\arguments{
\item{synID}{The Synapse ID of a table to query from. Defaults to
"syn10242922"}

\item{syn}{Synapse client object}
}
\value{
Data frame of table contents
}
\description{
Get the contents of a Synapse table as a data frame
}
\examples{
\dontrun{
syn <- synapse$Synapse()
syn$login()
get_synapse_table(synID = "syn10242922", syn = syn)
}
}
