% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-template.R
\name{get_template_keys_schema}
\alias{get_template_keys_schema}
\title{Get JSON schema template}
\usage{
get_template_keys_schema(syn, id = NA, file = NA)
}
\arguments{
\item{syn}{Synapse client object}

\item{id}{Registered synapse ID of schema.}

\item{file}{Filepath or URL to json schema.}
}
\description{
Get JSON schema registered in Synapse.
}
