% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-template.R
\name{get_template_keys_synID}
\alias{get_template_keys_synID}
\title{Get excel or csv template}
\usage{
get_template_keys_synID(syn, synID, ...)
}
\arguments{
\item{syn}{Synapse client object}

\item{synID}{Synapse ID of an excel or csv file containing a metadata
template}

\item{...}{Additional arguments passed to syn$get()}
}
\value{
Character vector of template column names
}
\description{
Get the metadata keys from an excel or csv template stored in
Synapse.
}
\examples{
\dontrun{
syn <- synapse$Synapse()
syn$login()
get_template("syn12973252", syn = syn)
}
}
