% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-template-dictionary.R
\name{get_template_synIDs}
\alias{get_template_synIDs}
\title{Get Template synIDs}
\usage{
get_template_synIDs(templates = get_golem_config("templates"))
}
\arguments{
\item{templates}{Named or unnamed list of template synIDs. Defaults to
\code{get_golem_config("templates")}.}
}
\value{
A vector of synIDs.
}
\description{
Get a vector of template synIDs from either a list or from the
config.yml file.
}
\examples{
\dontrun{
named_list <- dat1 <- list(
  template1 = "syn111111",
  template2 = "syn222222",
  template_set = list(
   template3 = "syn333333",
   template4 = "syn444444"
  )
)
get_template_synIDs(templates = named_list)

# config.yml example
# Write config file; this has fake synIDs and will not function
config <- list(
  default = list(
    templates = list(
      template1 = "syn222222",
      template2 = "syn3333333"
    )
  )
)
yaml::write_yaml(config, "./config.yml")
get_template_synIDs()
}
}
