% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-user-teams.R
\name{get_user_teams}
\alias{get_user_teams}
\title{Get teams a user belongs to}
\usage{
get_user_teams(user, syn)
}
\arguments{
\item{user}{Synapse user object (e.g. output from syn$getUserProfile())}

\item{syn}{Synapse client object}
}
\value{
Character vector of team IDs the user belongs to
}
\description{
Looks up the teams a user belongs to. By default, it looks for teams of the
current logged in user. You must be logged in to Synapse to use this
function.
}
\examples{
\dontrun{
syn <- synapse$Synapse()
syn$login()
user <- syn$getUserProfile()
get_user_teams(user, syn)
}
}
\keyword{internal}
