% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-synapse-oauth.R
\name{has_auth_code}
\alias{has_auth_code}
\title{Has authorization code}
\usage{
has_auth_code(params)
}
\arguments{
\item{params}{List of parsed URL parameters.}
}
\description{
Check if the authorization code is in the parsed URL parameters.
}
