% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-name.R
\name{is_name_valid}
\alias{is_name_valid}
\title{Checks that the name of the study or file is valid.
Valid strings contain: letters, numbers, spaces, underscores,
hyphens, periods, plus signs, and parentheses.}
\usage{
is_name_valid(name)
}
\arguments{
\item{name}{Name of file or study as a string.}
}
\value{
TRUE if name is valid; FALSE otherwise.
}
\description{
Checks that the name of the study or file is valid.
Valid strings contain: letters, numbers, spaces, underscores,
hyphens, periods, plus signs, and parentheses.
}
