% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-study-name.R
\name{is_study_name_valid}
\alias{is_study_name_valid}
\title{Checks that the name of the study is valid.}
\usage{
is_study_name_valid(study_name)
}
\arguments{
\item{study_name}{Name of the study.}
}
\value{
TRUE if study_name is valid; FALSE otherwise.
}
\description{
Checks that the name of the study is valid.
}
