% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse-helpers.R
\name{logged_in}
\alias{logged_in}
\title{Check if logged in as user}
\usage{
logged_in(syn)
}
\arguments{
\item{syn}{Synapse client object.}
}
\value{
FALSE if not logged in at all or if logged in anonymously, else TRUE.
}
\description{
Check if logged into Synapse as a non-anonymous user.
}
