% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_main_ui.R
\name{mod_main_server}
\alias{mod_main_server}
\title{Main dccvalidator server}
\usage{
mod_main_server(id, syn)
}
\arguments{
\item{id}{The module ID.}

\item{syn}{Synapse client object.}
}
\description{
Create the server-side component of the dccvalidator Shiny app.
}
