% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-main.R
\name{mod_main_ui}
\alias{mod_main_ui}
\alias{mod_main_server}
\title{Main dccvalidator UI}
\usage{
mod_main_ui(id)

mod_main_server(id, syn)
}
\arguments{
\item{id}{The module ID.}

\item{syn}{Synapse client object.}
}
\description{
Create the main UI and server components of the dccvalidator
Shiny app.
}
