% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-synapse-oauth.R
\name{mod_synapse_oauth_server}
\alias{mod_synapse_oauth_server}
\title{Synapse OAuth server}
\usage{
mod_synapse_oauth_server(id, syn)
}
\arguments{
\item{id}{The module id.}

\item{syn}{Synapse client object}
}
\description{
Provides functionality for the Synapse OAuth module. IMPORTANT:
this module assumes the following
global variables are available and valid: app, api, authorization_url,
app_url, claims_params, scope. See \code{\link{setup_global_oauth_vars}}.
}
