% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-synapse-oauth.R
\name{oauth_process}
\alias{oauth_process}
\title{Do the OAuth process}
\usage{
oauth_process(params)
}
\arguments{
\item{params}{List of parsed URL parameters.}
}
\value{
Nothing if the params have an authorization code already present.
Otherwise, kicks off the OAuth process and returns an access token for
log in.
}
\description{
Go through the OAuth process to get an access token for log in.
}
