% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse-oauth.R
\name{oauth_ui}
\alias{oauth_ui}
\title{OAuth UI}
\usage{
oauth_ui(request)
}
\arguments{
\item{request}{Shiny request object}
}
\description{
Launches the OAuth UI. If this is the first time signing in,
will need to do OAuth the process. The OAuth process will redirect back here
and an authorization code should be in the URL parameters. If this code is
available, lauch the main app_ui.
}
