% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-reporting.R
\name{report_result}
\alias{report_result}
\alias{report_results}
\title{Report results of validation checking}
\usage{
report_result(result, emoji_prefix = NULL, verbose = FALSE)

report_results(results, ...)
}
\arguments{
\item{result}{An object of type \code{check_pass}, \code{check_warn}, or \code{check_fail}
to report}

\item{emoji_prefix}{Emoji prefix to accompany reported result}

\item{verbose}{If \code{TRUE}, provides additional data about the check result
(accessed from the \code{$data} slot of the check object)}

\item{results}{A list of condition objects}

\item{...}{Additional parameters passed to \code{\link[=report_result]{report_result()}}}
}
\value{
A set of HTML elements showing the result of the check
}
\description{
Function to report results of a check. Shows an emoji (green check, yellow
warning, or red X), followed by the check message. If \code{verbose == TRUE},
includes a details draw that shows the invalid data. Each result has an
informational button to the right that users can click to learn more about
what the check was looking for (in the check object this message is stored in
"behavior").
}
\keyword{internal}
