% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-results-boxes.R
\name{results_boxes_server}
\alias{results_boxes_server}
\title{Server function for results boxes module}
\usage{
results_boxes_server(input, output, session, results)
}
\arguments{
\item{input}{The input from \code{\link[shiny:callModule]{shiny::callModule()}}.}

\item{output}{The output from \code{\link[shiny:callModule]{shiny::callModule()}}.}

\item{session}{The session from \code{\link[shiny:callModule]{shiny::callModule()}}.}

\item{results}{List of the validation results. If \code{NULL},
box titles will be default strings (i.e. "Successess (0)");
otherwise, the boxes will be populated with the
results.}
}
\description{
This gives functionality to the results boxes module
UI, attaching titles and populating
the validation results.
}
