% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse-oauth.R
\name{set_staging_endpoints}
\alias{set_staging_endpoints}
\title{Set Synapse Staging Endpoints}
\usage{
set_staging_endpoints(syn)
}
\arguments{
\item{syn}{Synapse client object}
}
\description{
By default, the Synapse Python client points to the production
endpoints. This function will point the client to the staging endpoints for
testing.
}
