% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-synapse-oauth.R
\name{setup_global_oauth_vars}
\alias{setup_global_oauth_vars}
\title{Sets up global OAuth variables}
\usage{
setup_global_oauth_vars(
  app_url,
  client_name,
  client_id,
  client_secret,
  claims = list(family_name = NULL, given_name = NULL, userid = NULL, is_certified =
    NULL),
  scope = "openid view download modify"
)
}
\arguments{
\item{app_url}{Application URL to redirect back to after OAuth process.}

\item{client_name}{Synapse OAuth client name.}

\item{client_id}{Synapse OAuth client ID.}

\item{client_secret}{Synapse OAuth client secret.}

\item{claims}{List of items to request access to. Defaults to \code{family_name},
\code{given_name}, \code{userid}, and \code{is_certified}. See ##TODO}

\item{scope}{Space-seperated access scope. Defaults to
\verb{openid view download modify}. See ##TODO}
}
\description{
Sets up the global OAuth variables needed for all users. These
must be initiated in the global environment for this to function correctly.
}
