% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-reporting.R
\name{show_details}
\alias{show_details}
\alias{show_details.default}
\alias{show_details.list}
\title{Show additional details in a drawer}
\usage{
show_details(x)

\method{show_details}{default}(x)

\method{show_details}{list}(x)
}
\arguments{
\item{x}{Content to be displayed}
}
\description{
S3 method to show details from a condition object. If the data included in
the object is a vector, it prints the elements separated by commas. If the
data is a list, it creates an HTML table with two columns: one containing the
list elements' names, and the other containing comma-separated values.
}
\keyword{internal}
