% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-helpers.R
\name{show_modal}
\alias{show_modal}
\title{Wrapper for modals}
\usage{
show_modal(title, message)
}
\arguments{
\item{title}{Title text for modal}

\item{message}{The message to show in the modal}
}
\description{
This wraps the functions for showing a modal with a message.
}
