% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{skip_on_fork}
\alias{skip_on_fork}
\title{Skip running a test on Travis from a forked repo}
\usage{
skip_on_fork()
}
\description{
Forked repos don't have access to the encrypted environment variables Travis
uses to run some tests. This function can be used to skip tests on builds
from forks.
}
