% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-template-dictionary.R
\name{update_template_dictionaries}
\alias{update_template_dictionaries}
\title{Update Template Dictionaries}
\usage{
update_template_dictionaries(..., active_config = "default")
}
\arguments{
\item{...}{Login parameters for the synapseclient. If none, assumes there is
a configuration file in the home directory with login information.
Optional parameters include combinations of \code{email} plus
\code{password} OR \code{email} plus \code{apiKey} OR \code{sessionToken}. \code{email} can be the
synapse username or the email address associated with the Synapse account.}

\item{active_config}{The configuration profile that should be active. The
config will be used to gather the annotation table synIDs and a list of
existing metadata template synIDs. Due to this, the config requires a minimum
of \code{annotations_table} and \code{templates}.}
}
\description{
This function acts as a script to update current metadata
template dictionaries. The metadata template will be an Excel file with three
sheets, template, dictionary, values. The template only has the required
column headers and all other rows are blank. This script assumes this part
exists within the Excel file. The dictionary sheet has two columns, key and
description, which has all the unique keys present in the template sheet and
their respective definitions. The values sheet has four columns, key, value,
valueDescription, and source. The keys are limited to those present in the
template sheet and values sheet contains the possible enumerated values for
those keys.
}
\examples{
\dontrun{
# Need a config file, this has fake synIDs and will not function
config <- list(
  default = list(
    annotations_table = "syn123456",
    templates = c("syn222222", "syn3333333")
  )
)
yaml::write_yaml(config, "./config.yml")
update_template_dictionaries(active_config = "default")
}
}
