% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-upload-documentation.R
\name{upload_docs_instruct_text}
\alias{upload_docs_instruct_text}
\title{Create instructions for uploading documentation}
\usage{
upload_docs_instruct_text(study_link_human, study_link_animal,
  study_link_ref)
}
\arguments{
\item{study_link_human}{html link to example of a study using human data}

\item{study_link_animal}{html link to example of a study using animal models}

\item{study_link_ref}{html link to example of study acknowledgment
preference}
}
\value{
html taglist with instructions
}
\description{
This function creates the instruction text for the Documentation
portion of dccvalidator. The study_link_ref argument dictates
whether instructions for uploading an acknowledgment is added.
If study_link_ref is an empty string, then the acknowledgement
instructions are not added.
}
