% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-upload-documentation.R
\name{upload_documents}
\alias{upload_documents}
\title{Server function for upload_documentation module}
\usage{
upload_documents(input, output, session, parent_folder)
}
\arguments{
\item{input}{the input from shiny::callModule}

\item{output}{the output from shiny::callModule}

\item{session}{the session from shiny::callModule}
}
\description{
Server function for upload_documentation module
}
