% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-upload-documentation.R
\name{upload_documents_server}
\alias{upload_documents_server}
\title{Server function for upload_documentation module}
\usage{
upload_documents_server(input, output, session, parent_folder,
  study_table_id)
}
\arguments{
\item{input}{the input from \code{\link[shiny:callModule]{shiny::callModule()}}}

\item{output}{the output from \code{\link[shiny:callModule]{shiny::callModule()}}}

\item{session}{the session from \code{\link[shiny:callModule]{shiny::callModule()}}}

\item{parent_folder}{the Synapse folder to put a Documentation folder in}

\item{study_table_id}{synapse Id for the consortium study table}
}
\description{
Server function for upload_documentation module
}
