% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-upload-documentation.R
\name{upload_documents_ui}
\alias{upload_documents_ui}
\title{UI function for the upload_documentation module}
\usage{
upload_documents_ui(id, study_link_human, study_link_animal)
}
\arguments{
\item{id}{the module id}

\item{study_link_human}{html link to example of a study using human data}

\item{study_link_animal}{html link to example of a study using animal models}
}
\value{
html ui for the module
}
\description{
UI function for the upload_documentation module
}
