% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-annotation-keys.R
\name{valid_annotation_keys}
\alias{valid_annotation_keys}
\title{Valid annotation keys}
\usage{
valid_annotation_keys(x, annotations, syn, ...)
}
\arguments{
\item{x}{An object to check.}

\item{annotations}{A data frame of annotation definitions. Must contain at
least three columns: \code{key}, \code{value}, and \code{columnType}.}

\item{syn}{Synapse client object}

\item{...}{Additional parameters passed to \code{\link[=check_keys]{check_keys()}}}
}
\value{
A vector of valid annotation keys present in \code{x}.
}
\description{
Checks for and returns the valid annotation keys in a data framae, Synapse
file, or Synapse file view.
}
\examples{
annots <- data.frame(
  key = c("assay", "fileFormat", "fileFormat", "fileFormat", "species"),
  value = c("rnaSeq", "fastq", "txt", "csv", "Human"),
  columnType = c("STRING", "STRING", "STRING", "STRING", "STRING")
)
dat1 <- data.frame(x = 1)
dat2 <- data.frame(assay = "rnaSeq")
valid_annotation_keys(dat1, annots)
valid_annotation_keys(dat2, annots)
}
