% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-template-dictionary.R
\name{verify_dictionary_structure}
\alias{verify_dictionary_structure}
\title{Verify Dictionary Structure}
\usage{
verify_dictionary_structure(dictionary)
}
\arguments{
\item{dictionary}{Dataframe with columns 'key', 'description',
and 'columnType'.}
}
\description{
Checks that no key has more than one description
or columnType. Throws error if the dictionary is invalid.
}
\examples{
# Valid
dat <- data.frame(
  key = c("first_name", "first_name"),
  description = c(
    "First name of individual",
    "First name of individual"
  ),
  columnType = c("string", "string")
)
verify_dictionary_structure(dat)

\dontrun{
# Invalid -- throws error due to descriptions of first_name
dat <- data.frame(
  key = c("first_name", "first_name"),
  description = c(
    "First name of individual",
    "Last name of individual"
  ),
  columnType = c("string", "string")
)
verify_dictionary_structure(dat)
}
}
