% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-button-indicator.R
\docType{data}
\name{with_busy_indicator_css}
\alias{with_busy_indicator_css}
\title{These functions add button feedback features including:
disabling the button while processing requested function,
showing a spinning wheel while processing requested function,
displaying a green checkmark showing success upon completion,
or displaying an error message if the function requested failed.}
\format{An object of class \code{character} of length 1.}
\usage{
with_busy_indicator_css
}
\description{
Author: Dean Attali
From https://github.com/daattali/advanced-shiny/blob/master/busy-indicator/helpers.R #nolint
}
\details{
Hint for making this work with modules by mmoise in PR#11
}
\keyword{datasets}
