% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-button-indicator.R
\name{with_busy_indicator_server}
\alias{with_busy_indicator_server}
\title{When a button is clicked, call this function from the server.
This will execute the expression desired and signal the visual
feedback for the button. If the expression fails, the error
message displayed is from the error that was thrown.}
\usage{
with_busy_indicator_server(button_id, expr)
}
\arguments{
\item{button_id}{id of shiny actionButton}

\item{expr}{the code to run when the button is clicked}
}
\description{
When a button is clicked, call this function from the server.
This will execute the expression desired and signal the visual
feedback for the button. If the expression fails, the error
message displayed is from the error that was thrown.
}
