% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-ui.R
\name{app_ui}
\alias{app_ui}
\title{App UI}
\usage{
app_ui(request)
}
\arguments{
\item{request}{Shiny request}
}
\value{
A shinydashboard page
}
\description{
Create the UI component of the dccvalidator Shiny app.
}
\examples{
\dontrun{
shinyApp(ui = app_ui, server = app_server)
}
}
