% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-annotation-values.R
\name{check_annotation_values}
\alias{check_annotation_values}
\title{Check annotation values}
\usage{
check_annotation_values(x, annotations, whitelist_keys = NULL,
  whitelist_values = NULL)
}
\arguments{
\item{x}{An object to check.}

\item{annotations}{A data frame of annotation definitions. Must contain at
least three columns: \code{key}, \code{value}, and \code{columnType}.}

\item{whitelist_keys}{A character vector keys to whitelist. If these keys are
present in \code{x} but absent from \code{annotations}, they will still be treated as
valid.}

\item{whitelist_values}{A named list of keys (as the names) and values (as
vectors) to whitelist}
}
\value{
A condition object indicating whether all annotation values are
valid. Invalid annotation values are included as data within the object.
}
\description{
Checks that all annotation values are valid. It does not report on values for
invalid \emph{keys}; see \code{\link[=check_annotation_keys]{check_annotation_keys()}}.
}
\examples{
\dontrun{
library("synapser")
synLogin()
annots <- get_synapse_annotations()
my_file <- synGet("syn11931757", downloadFile = FALSE)
check_annotation_values(my_file, annots)

dat <- data.frame(
  non_annotation = 5:7,
  assay = c("rnaSeq", "foo", "bar"),
  stringsAsFactors = FALSE
)
check_annotation_values(dat, annots)

fv <- synTableQuery("SELECT * FROM syn17020234")
check_annotation_values(fv, annots)

# If you don't specify an annotations data frame, these functions will
# download annotations automatically using `get_synapse_annotations()` (must
# be logged in to Synapse)
my_file <- synGet("syn11931757", downloadFile = FALSE)
check_annotation_values(my_file)

# It is possible to whitelist certain certain values, or all values for
# certain keys:
check_annotation_values(dat, whitelist_keys = "assay")
check_annotation_values(dat, whitelist_values = list(assay = c("foo")))
}
}
\seealso{
\code{\link[=valid_annotation_values]{valid_annotation_values()}}
}
