% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-col-names.R
\name{check_col_names}
\alias{check_col_names}
\alias{check_cols_manifest}
\alias{check_cols_individual}
\alias{check_cols_assay}
\alias{check_cols_biospecimen}
\title{Check column names against their corresponding template}
\usage{
check_col_names(data, template, success_msg = NULL, fail_msg = NULL,
  behavior = NULL)

check_cols_manifest(data, success_msg = "All manifest columns present",
  fail_msg = "Missing columns in the manifest", ...)

check_cols_individual(data, template,
  success_msg = "All individual metadata columns present",
  fail_msg = "Missing columns in the individual metadata file", ...)

check_cols_assay(data, template,
  success_msg = "All assay metadata columns present",
  fail_msg = "Missing columns in the assay metadata file", ...)

check_cols_biospecimen(data, template,
  success_msg = "All biospecimen columns present",
  fail_msg = "Missing columns in the biospecimen metadata file", ...)
}
\arguments{
\item{data}{Data file (manifest, individual metadata, or assay metadata)}

\item{template}{Character vector of column names from the template to check
against}

\item{success_msg}{Message indicating the check succeeded.}

\item{fail_msg}{Message indicating the check failed.}

\item{behavior}{The intended behavior of the test}

\item{...}{Additional arguments passed to \code{\link[synapser:synGet]{synapser::synGet()}}}
}
\value{
A condition object indicating whether the required columns were
present (\code{"check_pass"}) or absent (\code{"check_fail"}).
}
\description{
Check column names against their corresponding template
}
\examples{

a <- data.frame(path = "/path/file.txt", parent = "syn123", assay = "rnaSeq")
check_cols_manifest(a)

b <- data.frame(assay = "rnaSeq")
check_cols_manifest(b)
}
\seealso{
\code{\link[dccvalidator:get_template]{dccvalidator::get_template()}}
}
