% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-col-names.R
\name{check_col_names}
\alias{check_col_names}
\alias{check_cols_manifest}
\alias{check_cols_individual}
\alias{check_cols_assay}
\alias{check_cols_biospecimen}
\title{Check column names against their corresponding template}
\usage{
check_col_names(data, template)

check_cols_manifest(data)

check_cols_individual(data, template)

check_cols_assay(data, template)

check_cols_biospecimen(data)
}
\arguments{
\item{data}{Data file (manifest, individual metadata, or assay metadata)}

\item{template}{Character vector of column names from the template to check
against}
}
\value{
A character vector containing any column names that are missing from
the data file.
}
\description{
Check column names against their corresponding template
}
\seealso{
\code{\link[dccvalidator:get_template]{dccvalidator::get_template()}}
}
