% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-ids.R
\name{check_ids}
\alias{check_ids}
\alias{check_indiv_ids}
\alias{check_specimen_ids}
\title{Check ids}
\usage{
check_ids(x, y, idcol = c("individualID", "specimenID"), xname = NULL,
  yname = NULL)

check_indiv_ids(x, y, xname = NULL, yname = NULL)

check_specimen_ids(x, y, xname = NULL, yname = NULL)
}
\arguments{
\item{x, y}{Data frames to compare}

\item{idcol}{Name of column containing ids to compare}

\item{xname, yname}{Names of x and y (to be used in resulting messages)}
}
\value{
List of IDs missing from x (but present in y) and missing from y (but
present in x)
}
\description{
Compare IDs (such as individual IDs or specimen IDs) between two data frames.

Ensure that all individual IDs in two data frames match.

Ensure that all specimen IDS in two data frames match
}
