% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-ids.R
\name{check_indiv_ids}
\alias{check_indiv_ids}
\title{Check individual IDs}
\usage{
check_indiv_ids(x, y)
}
\arguments{
\item{x, y}{Data frames of metadata to compare}
}
\value{
A list of individual IDs missing from x (but present in y) and
missing from y (but present in x)
}
\description{
Ensure that all individual IDs in two data frames match.
}
