% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-ids.R
\name{check_indiv_ids}
\alias{check_indiv_ids}
\title{Check individual IDs}
\usage{
check_indiv_ids(clinical, assay)
}
\arguments{
\item{clinical}{Data frame of clinical metadata}

\item{assay}{Data frame of assay metadata}
}
\description{
Ensure that all individual IDs in assay metadata are present in clinical
metadata and vice versa.
}
