% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-check-team-membership.R
\name{check_team_membership}
\alias{check_team_membership}
\title{Check team membership}
\usage{
check_team_membership(input, output, session, team, user_teams)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}

\item{team}{Team ID to check membership in}

\item{user_teams}{Teams the user is a member of}
}
\description{
Shiny module to create a modal dialog.
}
