% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-team-membership.R
\name{check_team_membership}
\alias{check_team_membership}
\title{Check team membership}
\usage{
check_team_membership(teams, user)
}
\arguments{
\item{teams}{Team IDs to check membership in}

\item{user}{User to check (e.g. output from \code{\link[synapser:synGetUserProfile]{synapser::synGetUserProfile()}})}
}
\description{
Check if a user is a member of any of the given teams.
}
\keyword{internal}
