% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-annotation-values.R
\name{check_type}
\alias{check_type}
\alias{check_value}
\alias{check_values}
\title{Check that class of value matches annotation columnType}
\usage{
check_type(values, key, annotations, whitelist_values = NULL,
  return_valid = FALSE)

check_value(values, key, annotations, whitelist_keys = NULL,
  whitelist_values = NULL, return_valid = FALSE)

check_values(x, annotations, whitelist_keys = NULL,
  whitelist_values = NULL,
  success_msg = "All annotation values are valid",
  fail_msg = "Some annotation values are invalid",
  return_valid = FALSE)
}
\arguments{
\item{values}{The values of an annotation}

\item{key}{An annotation key}

\item{annotations}{A data frame of annotations to check against}

\item{whitelist_values}{A named list of keys (as the names) and values (as
vectors) to whitelist}

\item{return_valid}{Should the function return valid values? Defaults to
\code{FALSE} (i.e. the function will return invalid values).}

\item{whitelist_keys}{A character vector of annotation keys to whitelist. If
provided, all values for the given keys will be treated as valid.}

\item{x}{A data frame of annotation data}

\item{success_msg}{Message indicating the check succeeded.}

\item{fail_msg}{Message indicating the check failed.}
}
\value{
A vector of invalid values (if \code{return_valid = FALSE}; otherwise a
vector of valid values).

A character vector of valid or invalid values

If \code{return_valid = FALSE}: a condition object indicating whether all
annotation values are valid. Invalid annotation values are included as data
within the object: a named list where each element corresponds to a key
that contains invalid values, and the contents of each element is a vector
of invalid values. If \code{return_valid = TRUE}: a named list of the valid
annotation keys and values.
}
\description{
Given a vector of values, checks that they match the columnType as defined in
the annotations dictionary. This check is somewhat permissive in that values
that are coercible to the type are also treated as valid (see
\code{\link[=can_coerce]{can_coerce()}}).
}
\seealso{
\code{\link[=can_coerce]{can_coerce()}}
}
\keyword{internal}
