% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-folder.R
\name{create_folder}
\alias{create_folder}
\title{Create folder on Synapse}
\usage{
create_folder(parent, name)
}
\arguments{
\item{parent}{Synapse ID of the parent where the folder should be created}

\item{name}{Name of the folder to be created}
}
\value{
Synapse object of the folder that was created
}
\description{
Creates and stores a folder on Synapse under the given parent.
}
\examples{
\dontrun{
create_folder(parent = "syn17038062", name = "my_new_folder")
}
}
\keyword{internal}
