% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dccvalidator-package.R
\docType{package}
\name{dccvalidator-package}
\alias{dccvalidator}
\alias{dccvalidator-package}
\title{dccvalidator: Metadata Validation for Data Coordinating Centers}
\description{
Performs checks for common metadata quality issues. Used by the
    data coordinating centers for the 'AMP-AD' consortium
    (<https://adknowledgeportal.synapse.org>), 'PsychENCODE' consortium
    (<http://www.psychencode.org>), and others to validate metadata prior to
    data releases.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sage-bionetworks.github.io/dccvalidator}
  \item \url{https://github.com/Sage-Bionetworks/dccvalidator}
}

}
\author{
\strong{Maintainer}: Kara Woo \email{kara.woo@sagebase.org}

Authors:
\itemize{
  \item Nicole Kauer \email{nicole.kauer@sagebase.org}
  \item Kelsey Montgomery \email{kelsey.montgomery@sagebase.org}
}

Other contributors:
\itemize{
  \item Sage Bionetworks [copyright holder]
  \item Dean Attali [copyright holder]
}

}
\keyword{internal}
