% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df-helpers.R
\name{gather_template_ids}
\alias{gather_template_ids}
\title{Gather all template ids}
\usage{
gather_template_ids(type, species = NA, assay = NA)
}
\arguments{
\item{type}{The type of metadata template to get: manifest, individual,
biospecimen, or assay.}

\item{species}{The species needed to specify the correct biospecimen
or individual templates (default \code{NA}).}

\item{assay}{The assay needed to specify the correct assay template.}
}
\value{
the template id from the config (default \code{NA}).
}
\description{
Gather all template ids from config and add as a new column.
}
