% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-get-study-names.R
\name{get_study_names}
\alias{get_study_names}
\alias{get_study_ui}
\alias{get_study_server}
\title{Gather studies within consortium}
\usage{
get_study_names(study_table_id)

get_study_ui(id)

get_study_server(input, output, session, study_table_id)
}
\arguments{
\item{study_table_id}{synapse Id for the consortium study table}

\item{id}{the id}

\item{input}{the input variables from \code{\link[shiny:callModule]{shiny::callModule()}}}

\item{output}{the output variables from \code{\link[shiny:callModule]{shiny::callModule()}}}

\item{session}{the session from \code{\link[shiny:callModule]{shiny::callModule()}}}

\item{study_table_id}{synapse Id for the consortium study table}
}
\value{
list of studies

html UI for get_study panel

name of the study
}
\description{
Get study name based on either:
\itemize{
\item existing study selection box
\item new name for currently non-existing study
}
}
\keyword{internal}
