% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-template.R
\name{get_template_keys_schema}
\alias{get_template_keys_schema}
\title{Get JSON schema template}
\usage{
get_template_keys_schema(syn, id)
}
\arguments{
\item{syn}{Synapse client object}

\item{id}{The id of JSON metadata schema registered in Synapse or Synapse ID
of an excel or csv file containing a metadata template}
}
\description{
Get JSON schema registered in Synapse.
}
