% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-user-teams.R
\name{get_user_teams}
\alias{get_user_teams}
\title{Get teams a user belongs to}
\usage{
get_user_teams(user)
}
\arguments{
\item{user}{Synapser user object (e.g. output from
\code{\link[synapser:synGetUserProfile]{synapser::synGetUserProfile()}})}
}
\value{
Character vector of team IDs the user belongs to
}
\description{
Looks up the teams a user belongs to. By default, it looks for teams of the
current logged in user. You must be logged in to Synapse to use this
function.
}
\examples{
\dontrun{
user <- synapser::synGetUserProfile()
get_user_teams(user)
}
}
\keyword{internal}
