% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructions.R
\name{instructions}
\alias{instructions}
\title{Create instructions modal text}
\usage{
instructions(annots_link, templates_link)
}
\arguments{
\item{annots_link}{Link to a definition of the annotations being used in the
project}

\item{templates_link}{Link to the templates used for validation}
}
\value{
A div containin instructions
}
\description{
Provides an overview of the application and how to use it, as well as links
to resources about formatting a manifest and the allowable values of
annotations.
}
\keyword{internal}
