% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_main_ui.R
\name{mod_main_ui}
\alias{mod_main_ui}
\title{Main dccvalidator UI}
\usage{
mod_main_ui(id)
}
\arguments{
\item{id}{The module id.}
}
\description{
Create the main UI component of the dccvalidator Shiny app.
}
