% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-results-boxes.R
\name{results_boxes_ui}
\alias{results_boxes_ui}
\title{UI function for results boxes module}
\usage{
results_boxes_ui(id)
}
\arguments{
\item{id}{The module id.}
}
\value{
The html UI for the module.
}
\description{
This function outputs the html tags needed to
create the results boxes UI.
}
