% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny application}
\usage{
run_app(...)
}
\arguments{
\item{...}{Additional golem options passed to \code{\link[golem:with_golem_options]{golem::with_golem_options()}}}
}
\description{
Run the Shiny application
}
