% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-upload-documentation.R
\name{upload_documents_server}
\alias{upload_documents_server}
\title{Server function for upload_documentation module}
\usage{
upload_documents_server(input, output, session, parent_folder)
}
\arguments{
\item{input}{the input from \code{\link[shiny:callModule]{shiny::callModule()}}}

\item{output}{the output from \code{\link[shiny:callModule]{shiny::callModule()}}}

\item{session}{the session from \code{\link[shiny:callModule]{shiny::callModule()}}}
}
\description{
Server function for upload_documentation module
}
