% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-upload-documentation.R
\name{upload_documents_ui}
\alias{upload_documents_ui}
\title{UI function for the upload_documentation module}
\usage{
upload_documents_ui(id, study_table_id)
}
\arguments{
\item{id}{the module id}

\item{study_table_id}{synapse Id for the consortium study table}
}
\value{
html ui for the module
}
\description{
UI function for the upload_documentation module
}
