% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-annotation-values.R
\name{valid_annotation_values}
\alias{valid_annotation_values}
\title{Valid annotation values}
\usage{
valid_annotation_values(x, annotations, whitelist_keys = NULL,
  whitelist_values = NULL)
}
\arguments{
\item{x}{An object to check.}

\item{annotations}{A data frame of annotation definitions. Must contain at
least three columns: \code{key}, \code{value}, and \code{columnType}.}

\item{whitelist_keys}{A character vector keys to whitelist. If these keys are
present in \code{x} but absent from \code{annotations}, they will still be treated as
valid.}

\item{whitelist_values}{A named list of keys (as the names) and values (as
vectors) to whitelist}
}
\value{
A named list of valid annotation values.
}
\description{
Checks for and returns the valid annotation valaues in a data frame, Synapse
file, or Synapse file view.
}
