% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-annotation-values.R
\name{valid_annotation_values}
\alias{valid_annotation_values}
\title{Valid annotation values}
\usage{
valid_annotation_values(x, annotations, ...)
}
\arguments{
\item{x}{An object to check.}

\item{annotations}{A data frame of annotation definitions. Must contain at
least three columns: \code{key}, \code{value}, and \code{columnType}.}

\item{...}{Additional options to \code{\link[=check_values]{check_values()}}}

\item{syn}{Synapse client object}
}
\value{
A named list of valid annotation values.
}
\description{
Checks for and returns the valid annotation valaues in a data frame, Synapse
file, or Synapse file view.
}
