% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-button-indicator.R
\name{with_busy_indicator_ui}
\alias{with_busy_indicator_ui}
\title{Wrap the button in this function to attach visual features}
\usage{
with_busy_indicator_ui(button)
}
\arguments{
\item{button}{a shiny actionButton}
}
\description{
Wrap the button in this function to attach visual features
}
